refactor: workout
